+++ engim-app/engim-app/Models/Shapes/Position.cs
@@ -8,8 +8,8 @@ namespace engim_app.Models.Shapes
 {
     internal class Position
     {
-        public double X { get; set; }
-        public double Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
 
         public Position()
         {
